added ClimateActions context functionality
